# src/main/resources/swagger/swagger.yml
...
openapi: 3.0.0
info:
  title: Community REST API
  description: Community REST API
  version: "1.0"
...
servers:
  - url: http://localhost:8080/
  - url: https://localhost:8080/
################################################################################
#                               Components                                     #
################################################################################
components:
...
  schemas:
...
    NotificationType:
      type: string
      enum:
        - all
        - custom
...
    TopicType:
      type: string
      enum:
        - general
        - problem
        - system
        - eerpmGeneral
        - eerpmSummary
        - eerpmManual
        - eerpqGeneral
        - eerppGeneral
        - eerpmHighLevel
...
    UserStatus:
      type: string
      enum:
        - active
        - inactive
      default: active
...
    EerpqCodeType:
      type: string
      enum:
        - phenomenonCode
        - failureCode
        - dutyCode
        - reasonCode
        - solutionCode
...
    EerpType:
      type: string
      enum:
        - p
        - q
        - m
...
    SimpleGroupWithUsers:
      type: object
      required:
        - value
        - label
      properties:
        value:
          type: string
        label:
          type: string
        orgGroupType:
          type: string
        users:
          type: array
          items:
            $ref: '#/components/schemas/User'
...
    User:
      type: object
      properties:
        id:
          type: string
          default: ""
        name:
          type: string
          default: ""
        mail:
          type: string
          default: ""
        lock:
          type: boolean
          default: false
        roleId:
          type: integer
          format: int32
          default: 0
        totalCount:
          type: integer
          format: int32
          default: 0
        status:
          $ref: "#/components/schemas/UserStatus"
...
    TopicCreationData:
      allOf:
        - $ref: "#/components/schemas/ExtraRecipient"
      type: object
      required:
        - forumId
        - title
        - type
        - appField
      properties:
        forumId:
          type: integer
          format: int32
        title:
          type: string
        tag:
          type: array
          items:
            type: string
        type:
          $ref: "#/components/schemas/TopicType"
        text:
          type: string
        attachment:
          type: array
          items:
            $ref: "#/components/schemas/AttachmentWithAuthor"
        notificationType:
          $ref: "#/components/schemas/NotificationType"
        recipient:
          type: array
          items:
            type: string
        appField:
          type: array
          items:
            $ref: "#/components/schemas/LabelValueDto"
    AttachmentWithAuthor:
      allOf:
        - $ref: "#/components/schemas/VideoLanguageDto"
        - type: object
          properties:
            id:
              type: string
            author:
              type: array
              items:
                type: string
              minItems: 1
            appField:
              type: array
              items:
                $ref: "#/components/schemas/LabelValueDto"
            recordType:
              type: string
...
    DeltaPointInfo:
      type: object
      required:
        - levelName
        - levelCount
      properties:
        levelName:
          type: string
        levelCount:
          type: number
          format: float
          default: 0
        details:
          type: array
          items:
            $ref: "#/components/schemas/DeltaPointDetailInfo"
    DeltaPointDetailInfo:
      type: object
      required:
        - actionName
        - actionCount
      properties:
        actionName:
          type: string
        actionCount:
          type: number
          format: float
          default: 0
    DeltaPointLineChart:
      type: object
      properties:
        result:
          type: array
          items:
            $ref: "#/components/schemas/DeltaPointLineChartDetail"
        start:
          type: integer
          format: int32
          default: 1
    DeltaPointLineChartDetail:
      type: object
      properties:
        levelName:
          type: string
        points:
          type: array
          items:
            $ref: "#/components/schemas/LineChartPoint"
    LineChartPoint:
      type: object
      properties:
        x:
          type: integer
          format: int32
          default: 0
        y:
          type: integer
          format: int32
          default: 0
...
    EerpTopicCreationData:
      allOf:
        - $ref: "#/components/schemas/EerpTopicData"
        - type: object
          required:
            - forumId
          properties:
            forumId:
              type: integer
              format: int32
            type:
              $ref: "#/components/schemas/TopicType"
    EerpTopicData:
      type: object
      required:
        - title
      properties:
        title:
          type: string
        tag:
          type: array
          items:
            type: string
        text:
          type: string
    EerpmErrorCodeDto:
      type: object
      required:
        - errorCode
        - errorDesc
      properties:
        errorCode:
          type: string
        errorDesc:
          type: string
    EerpmErrorCauseDto:
      type: object
      required:
        - causeCode
        - causeDesc
      properties:
        causeCode:
          type: string
        causeDesc:
          type: string
    EerpmErrorSolutionDto:
      type: object
      required:
        - solutionCode
        - originSolution
      properties:
        solutionCode:
          type: string
        originSolution:
          type: string
    EerpqCodeDto:
      type: object
      properties:
        count:
          type: integer
          format: int32
        data:
          type: array
          items:
            $ref: "#/components/schemas/KeyLabelDto"
    KeyLabelDto:
      type: object
      properties:
        key:
          type: string
          nullable: false
        label:
          type: string
    LabelValueDto:
      type: object
      properties:
        label:
          type: string
          nullable: false
        value:
          type: object
...
    VideoLanguageDto:
      type: object
      properties:
        videoLanguage:
          type: string
          default: ""
...
    ExtraRecipient:
      type: object
      properties:
        orgMembers:
          type: array
          items:
            $ref: "#/components/schemas/SimpleGroupWithUsers"
        bgbus:
          type: array
          items:
            $ref: "#/components/schemas/SimpleGroupWithUsers"
...
paths:
...
  /topic:
    post:
      tags:
        - Topic
      summary: Create a topic
      description: Create a topic in the forum
      operationId: createTopic
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TopicCreationData"
        required: true
      responses:
        201:
          description: Success
          content:
            application/json:
              schema:
                type: integer
                format: int32
        400:
          description: Invalid parameters
        404:
          description: Create Forum Error
        413:
          description: File size exceeded
...
  /deltaPoint/{userId}:
    get:
      tags:
        - DeltaPoint
      summary: Get Delta Point of the user
      description: Get Delta Point of the user
      operationId: getDeltaPointResult
      parameters:
        - name: userId
          in: path
          description: Id of the user
          required: true
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DeltaPointInfo"
        400:
          description: Invalid parameters
  /deltaPoint/lineChart/{userId}:
    get:
      tags:
        - DeltaPoint
      summary: Get Delta Point LineChart of the user
      description: Get Delta Point LineChart of the user
      operationId: getDeltaPointLineChart
      parameters:
        - name: userId
          in: path
          description: Id of the user
          required: true
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeltaPointLineChart"
        400:
          description: Invalid parameters
  /deltaPoint/detail/{userId}:
    get:
      tags:
        - DeltaPoint
      summary: Get Delta Point Detail of the user
      description: Get Delta Point Detail of the user
      operationId: getDeltaPointDetail
      parameters:
        - name: userId
          in: path
          description: Id of the user
          required: true
          schema:
            type: string
        - name: type
          in: query
          description: 
            type
              * thisWeek
              * thisMonth
              * threeMonths
          required: true
          schema:
            type: string
            enum:
              - thisWeek
              - thisMonth
              - threeMonths
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DeltaPointInfo"
        400:
          description: Invalid parameters
...
  /topic/eerp:
    post:
      tags:
        - Eerp
      summary: Create a eerp topic
      description: Create a eerp topic in the forum
      operationId: createEerpTopic
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EerpTopicCreationData"
        required: true
      responses:
        200:
          description: Success
          content:
            application/json:
              schema:
                type: integer
                format: int32
        400:
          description: Invalid parameters
        404:
          description: Create pqm Error
  /topic/eerp/{topicId}:
    put:
      tags:
        - Eerp
      summary: Update the eerp topic information
      operationId: updateEerpTopic
      parameters:
        - name: topicId
          in: path
          description: Id of the topic
          required: true
          schema:
            type: integer
            format: int32
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EerpTopicData"
        required: true
      responses:
        200:
          description: Success
        400:
          description: Invalid parameters
        404:
          description: Update failed
...
  /topic/eerpm/errorcode:
    get:
      tags:
        - Eerp
      summary: get eerpm ErrorCode by deviceModel
      operationId: getEerpmErrorCode
      parameters:
        - name: deviceModel
          in: query
          description: Device Model
          required: true
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EerpmErrorCodeDto"
        400:
          description: Invalid parameters
  /topic/eerpm/errorcause:
    get:
      tags:
        - Eerp
      summary: get eerpm ErrorCause by deviceModel and errorCode
      operationId: getEerpmErrorCause
      parameters:
        - name: deviceModel
          in: query
          description: Device Model
          required: true
          schema:
            type: string
        - name: errorCode
          in: query
          description: Error Code
          required: true
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EerpmErrorCauseDto"
        400:
          description: Invalid parameters
  /topic/eerpm/errorsolution:
    get:
      tags:
        - Eerp
      summary: get eerpm ErrorSolution by deviceModel, errorCode and errorCause
      operationId: getEerpmErrorSolution
      parameters:
        - name: deviceModel
          in: query
          description: Device Model
          required: true
          schema:
            type: string
        - name: errorCode
          in: query
          description: Error Code
          required: true
          schema:
            type: string
        - name: errorCause
          in: query
          description: Error Cause
          required: true
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EerpmErrorSolutionDto"
        400:
          description: Invalid parameters
  /topic/eerpq/code:
    get:
      tags:
        - Eerp
      summary: get eerpq code
      operationId: getEerpqCode
      parameters:
        - name: pageNum
          in: query
          schema:
            type: integer
            format: int32
            default: -1
        - name: pageSize
          in: query
          schema:
            type: integer
            format: int32
            default: -1
        - name: type
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/EerpqCodeType"
        - name: factory
          in: query
          required: true
          schema:
            type: string
        - name: phenomenonCode
          in: query
          schema:
            type: string
        - name: dutyCode
          in: query
          schema:
            type: string
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EerpqCodeDto"
        400:
          description: Invalid parameters
...
  /topic/eerp/report/conclusion:
    post:
      tags:
        - Eerp
      summary: generate conclusion report to MyDms
      operationId: generateEerpConclusionReport
      parameters:
        - name: type
          in: query
          required: true
          schema:
            $ref: "#/components/schemas/EerpType"
        - name: startTime
          in: query
          required: false
          schema:
            type: integer
            format: int64
            default: -1
        - name: endTime
          in: query
          required: true
          schema:
            type: integer
            format: int64
      responses:
        200:
          description: OK
        401:
          description: Not login
        403:
          description: Permission denied
  /dia/attachment:
    post:
      tags:
        - Dia
      summary: download attachments and create ddf of dia topics
      operationId: createAllDiaAttachment
      responses:
        200:
          description: OK
        401:
          description: Not login
        403:
          description: Permission denied
  /dia/topic:
    post:
      tags:
        - Dia
      summary: create all dia topics
      operationId: createAllDiaTopic
      responses:
        200:
          description: OK
        401:
          description: Not login
        403:
          description: Permission denied
...